+++ lib/src/model/form_decimal_item.dart
@@ -37,6 +37,4 @@ class FormDecimalItem extends FormTextItem {
       return null;
     }
   }
-
-
 }