+++ lib/src/model/form_int_item.dart
@@ -31,6 +31,4 @@ class FormIntItem extends FormTextItem {
       return null;
     }
   }
-
-
 }